Long 改为 Integer

+++ src/main/java/cn/staitech/annotation/vo/anno/AnnotationAddReq.java
@@ -92,5 +92,5 @@ public class AnnotationAddReq {
     private String jsonId;
 
     @ApiModelProperty(value = "脏器识别校对view页面必传参数：传递1")
-    private Long contourType;
+    private Integer contourType;
 }
